refactor(NcCounterBubble): migrate component to Typescript and script-setup
