+++ src/components/NcCounterBubble/index.ts
