+++ tests/unit/components/NcCounterBubble/NcCounterBubble.spec.ts
